refactor: 패키지 폴더 정리
