+++ src/main/java/domain/piece/Color.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.piece;
 
 import java.util.Arrays;
