+++ src/main/java/domain/piece/objects/Rook.java
@@ -1,7 +1,9 @@
-package domain.piece;
+package domain.piece.objects;
 
-import domain.Direction;
-import domain.Score;
+import domain.piece.Direction;
+import domain.piece.Piece;
+import domain.piece.Position;
+import domain.score.Score;
 
 import java.util.HashMap;
 import java.util.Map;
