+++ src/main/java/domain/score/Score.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.score;
 
 import java.util.*;
