+++ src/test/java/domain/piece/ColorTest.java
@@ -1,5 +1,6 @@
-package domain;
+package domain.piece;
 
+import domain.piece.Color;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
