+++ src/test/java/domain/piece/objects/QueenTest.java
@@ -1,8 +1,10 @@
-package domain.piece;
+package domain.piece.objects;
 
 import domain.Board;
-import domain.PieceFactory;
-import javafx.geometry.Pos;
+import domain.piece.Piece;
+import domain.piece.Position;
+import domain.piece.objects.Pawn;
+import domain.piece.objects.Queen;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
