+++ src/test/java/domain/piece/objects/RookTest.java
@@ -1,7 +1,11 @@
-package domain.piece;
+package domain.piece.objects;
 
 import domain.Board;
-import domain.PieceFactory;
+import domain.piece.Piece;
+import domain.piece.Position;
+import domain.piece.objects.Pawn;
+import domain.piece.objects.Queen;
+import domain.piece.objects.Rook;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
